% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKF.R
\name{optim_params}
\alias{optim_params}
\title{optim_params
Uses pracma optimization methods for either Simulated
Annealing or Nelder-Meade. Calls UKF_blend to apply UKF
to time series data and model to update parameters and return
chi-square goodness of fit for the objective function.}
\usage{
optim_params(
  param_guess,
  method = "L-BFGS-B",
  lower_lim,
  upper_lim,
  maxit,
  temp = 20,
  t_dummy,
  ts_data,
  ode_model,
  N_p,
  N_y,
  dt,
  dT,
  forcePositive = FALSE
)
}
\arguments{
\item{param_guess}{vector of length N_p with initial guess for ode_model parameters}

\item{method}{string to specify optimization method,"L-BFGS-B" or "SANN"}

\item{lower_lim}{lower bound of solution (L-BFGS only)}

\item{upper_lim}{upper bound of solution (L-BFGS only)}

\item{maxit}{maximum number of iterations (SANN only)}

\item{temp}{annealing temperature}

\item{t_dummy}{a dummy time variable, because ode models don't have explcit time}

\item{ts_data}{panel of time series data. First column holds time values, other N_y columns hold variable values}

\item{ode_model}{model function with ind variables y and N_p parameters}

\item{N_p}{number of unknown model params}

\item{N_y}{number of ind variables}

\item{dt}{smaller time step size within dT for solving ode}

\item{dT}{time step size that comes from time series data step}

\item{forcePositive}{logical, if TRUE, ensures all parameters stay positive}
}
\value{
list: par (vector of N_p optimized parameters) and value (final chi-square goodness of fit to time series)
}
\description{
optim_params
Uses pracma optimization methods for either Simulated
Annealing or Nelder-Meade. Calls UKF_blend to apply UKF
to time series data and model to update parameters and return
chi-square goodness of fit for the objective function.
}
\examples{
Example
}
