% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_smooth.R
\name{plot_ukf_and_smoothed}
\alias{plot_ukf_and_smoothed}
\title{plot_ukf_and_smoothed
Plot a pair of voxel time series and UKF blended data.
Panel for each voxel with input data and UKF estimate.
The legend says Kernel for input data, but could input raw.
So, might want to make legend text an input}
\usage{
plot_ukf_and_smoothed(
  ukf_out,
  smoothed_data,
  top_title = "UKF estimate and kernel smooth"
)
}
\arguments{
\item{ukf_out}{output object from UKF_blend}

\item{smoothed_data}{time series panel used for UKF_blend}

\item{top_title}{string title typically to describe how parameters were optimized}
}
\value{
Just a plot
}
\description{
plot_ukf_and_smoothed
Plot a pair of voxel time series and UKF blended data.
Panel for each voxel with input data and UKF estimate.
The legend says Kernel for input data, but could input raw.
So, might want to make legend text an input
}
\examples{
Example
}
