% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_smooth.R
\name{plot_ukf_and_smoothed_ggplot}
\alias{plot_ukf_and_smoothed_ggplot}
\title{plot_ukf_and_smoothed_ggplot
Plot UKF estimate and kernel smooth data.}
\usage{
plot_ukf_and_smoothed_ggplot(
  ukf_out,
  smoothed_data,
  top_title = "UKF estimate and kernel smooth"
)
}
\arguments{
\item{ukf_out}{Output from the UKF function.}

\item{smoothed_data}{Smoothed data.}

\item{top_title}{Title for the plot.}
}
\value{
None
}
\description{
plot_ukf_and_smoothed_ggplot
Plot UKF estimate and kernel smooth data.
}
