% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_smooth.R
\name{plot_ukf_chi_square_loss}
\alias{plot_ukf_chi_square_loss}
\title{Plot the chi-square loss from the UKF output.}
\description{
This function assumes that the UKF output contains a 'chisq' component.
}
\usage{
plot_ukf_chi_square_loss(ukf_output, title = "UKF Chi-Square Loss Over Time")
}
\arguments{
\item{ukf_output}{Output from the UKF function.}

\item{title}{Title for the plot.}
}
\value{
None
}
\examples{
# Example usage:
plot_ukf_chi_square_loss(ukf_output, title = "Chi-Square Loss Example")
}
\seealso{
\code{\link{ggplot2}}
}
