% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_smooth.R
\name{plot_voxels_and_smooth}
\alias{plot_voxels_and_smooth}
\title{plot_voxels_and_smooth
Plot a pair of voxel time series and kernel smoothed data.
Panel for each voxel with raw and smoothed.
Also returns the smoothed data for later use.}
\usage{
plot_voxels_and_smooth(voxel_input)
}
\arguments{
\item{voxel_input}{three columns: column 1 = time, column 2 = voxel 1 signal, column 3 = voxel 2 signal}
}
\value{
matrix: same shape as input data but kernel smoothed
}
\description{
plot_voxels_and_smooth
Plot a pair of voxel time series and kernel smoothed data.
Panel for each voxel with raw and smoothed.
Also returns the smoothed data for later use.
}
\examples{
Example
}
