% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKF.R
\name{UKF_dT}
\alias{UKF_dT}
\title{UKF_dT
Unscented Kalman Filter (UKF) for time step dT
Doesn't use data at this step;
UKF_blend fn later blends UKF with data after dT step.
This function was known previously as
multiple_param_unscented_kalman_filter.}
\usage{
UKF_dT(t_dummy, ode_model, xhat, Pxx, y, N_p, N_y, R, dt, dT, R_scale, Q_scale, forcePositive)
}
\arguments{
\item{t_dummy}{a dummy time variable, because ode models don't have explcit time}

\item{ode_model}{model with variables y and N_p parameters}

\item{xhat}{UFK posterior prediction of augmented states (param + ind vars)}

\item{Pxx}{augmented covariance matrix for sigma points}

\item{y}{independent variable state vector}

\item{N_p}{number of unknown model params}

\item{N_y}{number of ind variables}

\item{R}{covariance matrix of y state. initialized in UKF_blend and updated with sigma points as Pyy in current function.}

\item{dt}{smaller time step size within dT}

\item{dT}{time step size that comes from time series data step}

\item{R_scale}{related to standard deviation of Gaussian measurement noise of ind variables. A number that must be specified by user}

\item{Q_scale}{related to standard deviation of process noise. Noise related to model parameters. User choice. Can it be 0?}

\item{forcePositive}{logical, if TRUE, ensures all parameters stay positive}
}
\value{
list: xhat (augmented Kalman update), Pxx (augmented covariance at sigma points), and K (Kalman gain matrix)
}
\description{
UKF_dT
Unscented Kalman Filter (UKF) for time step dT
Doesn't use data at this step;
UKF_blend fn later blends UKF with data after dT step.
This function was known previously as
multiple_param_unscented_kalman_filter.
}
\examples{
Example
}
