% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKF.R
\name{iterative_param_optim}
\alias{iterative_param_optim}
\title{iterative_param_optim
Optimize the model parameters by iteratively running
the UKF_blend through the time series data.
In other words, repeated calls of UKF_blend to apply UKF
to time series data and model to update parameters and return
chi-square goodness of fit for the objective function.
Convergence of the parameter vector between runs is the
stopping criterion}
\usage{
iterative_param_optim(
  param_guess,
  t_dummy,
  ts_data,
  ode_model,
  N_p,
  N_y,
  dt,
  dT,
  param_tol = 0.01,
  MAXSTEPS = 30,
  forcePositive = FALSE
)
}
\arguments{
\item{param_guess}{vector of length N_p with initial guess for ode_model parameters}

\item{t_dummy}{a dummy time variable, because ode models don't have explcit time}

\item{ts_data}{panel of time series data. First column holds time values, other N_y columns hold variable values}

\item{ode_model}{model function with ind variables y and N_p parameters}

\item{N_p}{number of unknown model params}

\item{N_y}{number of ind variables}

\item{dt}{smaller time step size within dT for solving ode}

\item{dT}{time step size that comes from time series data step}

\item{param_tol}{small tolerance for convergence of the parameters to be optimized}

\item{MAXSTEPS}{max number of iterations through time series}

\item{forcePositive}{logical, if TRUE, ensures all parameters stay positive}
}
\value{
list: par (vector of N_p optimized parameters) and value (final chi-square goodness of fit to time series)
}
\description{
iterative_param_optim
Optimize the model parameters by iteratively running
the UKF_blend through the time series data.
In other words, repeated calls of UKF_blend to apply UKF
to time series data and model to update parameters and return
chi-square goodness of fit for the objective function.
Convergence of the parameter vector between runs is the
stopping criterion
}
\examples{
Example
}
