% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKF.R
\name{UKF_dT}
\alias{UKF_dT}
\title{UKF_dT
Unscented Kalman Filter (UKF) for time step dT
Doesn't use data yet. UKF_blend fn blends after dT.
This function was known previously as
multiple_param_unscented_kalman_filter.
Input: t_dummy,ode_model,xhat,Pxx,y,dq,dy,R,dt,dT}
\usage{
UKF_dT(t_dummy, ode_model, xhat, Pxx, y, dq, dy, R, dt, dT)
}
\arguments{
\item{t_dummy}{a dummy time variable, because ode models don't have explcit time}

\item{ode_model}{model with variables y and dq parameters}

\item{xhat}{UFK prediction of augmented states (param + ind vars)}

\item{Pxx}{augmented covariance matrix for sigma points}

\item{y}{independent variable state vector}

\item{dq}{number of unknown model params}

\item{dy}{number of ind variables}

\item{R}{covariance matrix of y state. initialized in UKF_blend and updated with sigma points as Pyy in current function.}

\item{dt}{smaller time step size within dT}

\item{dT}{time step size that comes from time series data step}
}
\value{
list: xhat (augmented Kalman update), Pxx (augmented covariance at sigma points), and K (Kalman gain matrix)
}
\description{
UKF_dT
Unscented Kalman Filter (UKF) for time step dT
Doesn't use data yet. UKF_blend fn blends after dT.
This function was known previously as
multiple_param_unscented_kalman_filter.
Input: t_dummy,ode_model,xhat,Pxx,y,dq,dy,R,dt,dT
}
\examples{
Example
}
